re-organization of notebook utils
